Move AssertionUtils to kyuubi-util-scala module
