+++ kyuubi-util-scala/src/test/java/org/apache/kyuubi/util/AssertionUtils.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kyuubi.plugin.spark.authz.util
+package org.apache.kyuubi.util
 
 import java.util.Locale
